refactor: modularize CombinedExpressionEvaluator into focused modules
Refactored the combined expression evaluator from a monolithic 437-line
file into a modular architecture with clear separation of concerns.

## Changes

### New Module Structure

Created `combined/` subdirectory with focused modules:
- **eval.rs**: Main dispatch logic and column references
- **predicates.rs**: BETWEEN, LIKE, IN, IS NULL predicates
- **subqueries.rs**: Scalar subqueries, EXISTS, quantified comparisons
- **special.rs**: CASE expressions, CAST, function calls

### Benefits

- **Reduced complexity**: 437-line file split into 4 modules of ~100-200 lines each
- **Single Responsibility**: Each module handles specific expression types
- **Maintainability**: Changes isolated to relevant modules
- **Extensibility**: Easy to add new expression types
- **Consistency**: Mirrors existing ExpressionEvaluator organization

### Documentation

Added `EVALUATOR_ARCHITECTURE.md` documenting:
- Module organization and responsibilities
- Design principles and shared logic
- Future trait-based enhancement path
- Migration guide (no breaking changes)

## Testing

✅ All existing tests pass without modification
✅ No performance impact (static dispatch maintained)
✅ Full API compatibility preserved

Closes #337
