+++ crates/executor/src/evaluator/combined/subqueries.rs
@@ -1,0 +1,216 @@
+///! Subquery evaluation for combined expressions
+
+use crate::errors::ExecutorError;
+use super::super::core::{CombinedExpressionEvaluator, ExpressionEvaluator};
+
+impl<'a> CombinedExpressionEvaluator<'a> {
+    /// Evaluate scalar subquery - must return exactly one row and one column
+    pub(super) fn eval_scalar_subquery(
+        &self,
+        subquery: &ast::SelectStmt,
+        row: &storage::Row,
+    ) -> Result<types::SqlValue, ExecutorError> {
+        let database = self.database.ok_or(ExecutorError::UnsupportedFeature(
+            "Subquery execution requires database reference".to_string(),
+        ))?;
+
+        // Execute the subquery using SelectExecutor
+        let select_executor = crate::select::SelectExecutor::new(database);
+        let rows = select_executor.execute(subquery)?;
+
+        // SQL:1999 Section 7.9: Scalar subquery must return exactly 1 row
+        if rows.len() > 1 {
+            return Err(ExecutorError::SubqueryReturnedMultipleRows {
+                expected: 1,
+                actual: rows.len(),
+            });
+        }
+
+        // SQL:1999 Section 7.9: Scalar subquery must return exactly 1 column
+        // Check column count from SELECT list
+        if subquery.select_list.len() != 1 {
+            return Err(ExecutorError::SubqueryColumnCountMismatch {
+                expected: 1,
+                actual: subquery.select_list.len(),
+            });
+        }
+
+        // Return the single value, or NULL if no rows
+        if rows.is_empty() {
+            Ok(types::SqlValue::Null)
+        } else {
+            rows[0]
+                .get(0)
+                .cloned()
+                .ok_or(ExecutorError::ColumnIndexOutOfBounds { index: 0 })
+        }
+    }
+
+    /// Evaluate EXISTS predicate: EXISTS (SELECT ...)
+    /// SQL:1999 Section 8.7: EXISTS predicate
+    /// Returns TRUE if subquery returns at least one row
+    /// Returns FALSE if subquery returns zero rows
+    /// Never returns NULL (unlike most predicates)
+    pub(super) fn eval_exists(
+        &self,
+        subquery: &ast::SelectStmt,
+        negated: bool,
+        row: &storage::Row,
+    ) -> Result<types::SqlValue, ExecutorError> {
+        let database = self.database.ok_or(ExecutorError::UnsupportedFeature(
+            "EXISTS requires database reference".to_string(),
+        ))?;
+
+        // Execute the subquery using SelectExecutor
+        let select_executor = crate::select::SelectExecutor::new(database);
+        let rows = select_executor.execute(subquery)?;
+
+        // Check if subquery returned any rows
+        let has_rows = !rows.is_empty();
+
+        // Apply negation if needed
+        let result = if negated { !has_rows } else { has_rows };
+
+        Ok(types::SqlValue::Boolean(result))
+    }
+
+    /// Evaluate quantified comparison: expr op ALL/ANY/SOME (SELECT ...)
+    /// SQL:1999 Section 8.8: Quantified comparison predicate
+    /// ALL: comparison must be TRUE for all rows
+    /// ANY/SOME: comparison must be TRUE for at least one row
+    pub(super) fn eval_quantified(
+        &self,
+        expr: &ast::Expression,
+        op: &ast::BinaryOperator,
+        quantifier: &ast::Quantifier,
+        subquery: &ast::SelectStmt,
+        row: &storage::Row,
+    ) -> Result<types::SqlValue, ExecutorError> {
+        let database = self.database.ok_or(ExecutorError::UnsupportedFeature(
+            "Quantified comparison requires database reference".to_string(),
+        ))?;
+
+        // Evaluate the left-hand expression
+        let left_val = self.eval(expr, row)?;
+
+        // Execute the subquery using SelectExecutor
+        let select_executor = crate::select::SelectExecutor::new(database);
+        let rows = select_executor.execute(subquery)?;
+
+        // Empty subquery special cases:
+        // - ALL: returns TRUE (vacuously true - all zero rows satisfy the condition)
+        // - ANY/SOME: returns FALSE (no rows to satisfy the condition)
+        if rows.is_empty() {
+            return Ok(types::SqlValue::Boolean(matches!(quantifier, ast::Quantifier::All)));
+        }
+
+        // If left value is NULL, result depends on quantifier and subquery results
+        if matches!(left_val, types::SqlValue::Null) {
+            return Ok(types::SqlValue::Null);
+        }
+
+        match quantifier {
+            ast::Quantifier::All => {
+                // ALL: comparison must be TRUE for all rows
+                // If any comparison is FALSE, return FALSE
+                // If any comparison is NULL (and none FALSE), return NULL
+                let mut has_null = false;
+
+                for subquery_row in &rows {
+                    if subquery_row.values.len() != 1 {
+                        return Err(ExecutorError::SubqueryColumnCountMismatch {
+                            expected: 1,
+                            actual: subquery_row.values.len(),
+                        });
+                    }
+
+                    let right_val = &subquery_row.values[0];
+
+                    // Handle NULL in subquery result
+                    if matches!(right_val, types::SqlValue::Null) {
+                        has_null = true;
+                        continue;
+                    }
+
+                    // Evaluate comparison
+                    let cmp_result = ExpressionEvaluator::eval_binary_op_static(&left_val, op, right_val)?;
+
+                    match cmp_result {
+                        types::SqlValue::Boolean(false) => return Ok(types::SqlValue::Boolean(false)),
+                        types::SqlValue::Null => has_null = true,
+                        _ => {} // TRUE, continue checking
+                    }
+                }
+
+                // If we saw any NULLs (and no FALSEs), return NULL
+                // Otherwise return TRUE (all comparisons were TRUE)
+                if has_null {
+                    Ok(types::SqlValue::Null)
+                } else {
+                    Ok(types::SqlValue::Boolean(true))
+                }
+            }
+
+            ast::Quantifier::Any | ast::Quantifier::Some => {
+                // ANY/SOME: comparison must be TRUE for at least one row
+                // If any comparison is TRUE, return TRUE
+                // If all comparisons are FALSE, return FALSE
+                // If any comparison is NULL (and none TRUE), return NULL
+                let mut has_null = false;
+
+                for subquery_row in &rows {
+                    if subquery_row.values.len() != 1 {
+                        return Err(ExecutorError::SubqueryColumnCountMismatch {
+                            expected: 1,
+                            actual: subquery_row.values.len(),
+                        });
+                    }
+
+                    let right_val = &subquery_row.values[0];
+
+                    // Handle NULL in subquery result
+                    if matches!(right_val, types::SqlValue::Null) {
+                        has_null = true;
+                        continue;
+                    }
+
+                    // Evaluate comparison
+                    let cmp_result = ExpressionEvaluator::eval_binary_op_static(&left_val, op, right_val)?;
+
+                    match cmp_result {
+                        types::SqlValue::Boolean(true) => return Ok(types::SqlValue::Boolean(true)),
+                        types::SqlValue::Null => has_null = true,
+                        _ => {} // FALSE, continue checking
+                    }
+                }
+
+                // If we saw any NULLs (and no TRUEs), return NULL
+                // Otherwise return FALSE (no comparisons were TRUE)
+                if has_null {
+                    Ok(types::SqlValue::Null)
+                } else {
+                    Ok(types::SqlValue::Boolean(false))
+                }
+            }
+        }
+    }
+
+    /// Evaluate IN operator with subquery
+    /// Currently returns an error as it requires database access
+    pub(super) fn eval_in_subquery(
+        &self,
+        expr: &ast::Expression,
+        _subquery: &ast::SelectStmt,
+        _negated: bool,
+        row: &storage::Row,
+    ) -> Result<types::SqlValue, ExecutorError> {
+        // TODO: Full implementation requires database access to execute subquery
+        // This requires refactoring CombinedExpressionEvaluator to have database reference
+        // For now, evaluate the left expression to ensure it's valid
+        let _left_val = self.eval(expr, row)?;
+        Err(ExecutorError::UnsupportedFeature(
+            "IN with subquery requires database access - implementation pending"
+                .to_string(),
+        ))
+    }
+}
